8. String to Integer (atoi)
